API: schemas: New crm_shadow and instruction schemas
The instruction schema currently exists only to support the crm_shadow
schema. It may find wider use later, so we give it its own file.

The shadow message formatter function doesn't care whether the instance
attribute is empty. But in practice, crm_shadow always populates the
instance attribute, so we require it in the schema.

Ref T108

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>