Test: cts-cli: Improve sed for cib_last_written and shadow directory
Avoid mangling lines like

<cib epoch="4" num_updates="1" admin_epoch="1"/>

and

<change-attr name="cib-last-written" operation="unset"/>

when we add XML formatted output for crm_shadow.

Also, match the shadow directory when it's an XML attribute instead of
only at the beginning of a line.

Note that these sed expressions are basically duplicated. This is for
POSIX compliance (specifically to avoid CI failures on FreeBSD). POSIX
sed doesn't support EREs, which includes '?' as a special character
meaning zero-or-one.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>