API: schemas: New diff and any-element schemas
The diff schema validates an XML patchset of format v2. Format v1 is
quite old and won't be added to the schema.

any-element is a helper that, as the name suggests, validates an XML
element with arbitrarily nested children.

There are no current plans to include these in other API schemas.
However, the diff schema is useful for specifying the expected format of
an XML patchset for anyone who may want to parse them, and it provides a
way to validate a diff element if we so choose.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>