API: libcrmcommon: New PCMK_META_ENABLED string constant
The new constant represents "enabled" and will be used soon to indicate
that a particular alert or recipient is enabled. In the future, it may
be used as a meta attribute for other entities besides alerts and
recipients.

This commit also introduces a new convention in the msg_xml.h comments:
use PCMK_META_ as a prefix for meta attribute names.

Ref T379

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>