Feature: libcrmcommon, libpe_status: New enabled meta attr for alerts
An <alert> element or an individual <recipient> element can now contain
a boolean "enabled" meta attribute. The default value is "true". If
"enabled" is set to "false", the alert or recipient will not be used.

If "enabled" is set to "false" at the <alert> level, all of the alert's
recipients are disabled, regardless of the recipients' "enabled" value.

If "enabled" is set to "true" or is unset at the <alert> level, the
alert's recipients are enabled or disabled based on their respective
values of the "enabled" meta attribute.

Setting "enabled" to "false" can be useful to avoid spamming alert
recipients during planned maintenance, for example.

Ref T379

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>