API: libcrmcommon: New pcmk_rc_disabled return code
This new return code is to be used when a particular object is disabled.
The motivating use case is unpacking a disabled alert (which is a new
feature that will appear in an upcoming commit).

It can be used for any disabled configuration object that we want to
ignore during unpacking. This is distinct from pcmk_rc_unpack_error,
which indicates that something is malformed or has an unsupported value.

More generally, this code can be used any time we need to know that
something (whether part of the configuration or not) is disabled and
there's no error condition otherwise.

Ref T379

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>