Feature: daemons: Add daemon uptime to execd XML replies.
When the control daemon connects to the executor, the response it gets
back now includes its uptime as an XML attribute.  The uptime is really
just how much real time has elapsed since the process was started.  It
doesn't take into account CPU time or anything of that nature.

This is added to the private remote-specific data in the lrmd_t.

Ref T133