Refactor: python: Disable warning about the getInstance name.
I haven't decided how to deal with this yet, but for the moment I at
least want pylint to not report any errors.  One possible fix is
renaming the function.  Another possible fix is converting the
RemoteExec class to a singleton and just using that directly, skipping
the RemoteFactory.  Python users overall are not big fans of singleton
classes, though, so I'm not sure I actually want to do this.