Refactor: tools: New enum mon_exec_mode for crm_mon
There are three mutually exclusive modes of execution: interactive
(console), one-shot, and daemonized. Currently we store booleans called
options.one_shot and options.daemonize, and when we set one to true, we
set the other to false. If both are false, then we use interactive mode.

It's less confusing and less error-prone to use a single enum variable
to represent the execution mode.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>