Feature: tools: crm_mon can use update mode with XML, text, none
This feature allows a user to keep a crm_mon process running in the
foreground while updating an output file at regular intervals, if the
"--interval" option is passed and "--daemonize" is not passed. The
behavior is much like a foreground version of --daemonize.

HTML output already allowed this, although this commit may address
corner cases. It's unclear whether other formats allowed this in the
past.

There's no output with "none" of course, but this allows the user to
keep a crm_mon process in the foreground and use --external-agent if
they want.

There may not be much of a use case for update mode for text. However,
the behavior is consistent this way.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>