Feature: controller: Remove purge flag when the scheduler says to.
When we see a start action for a remote connection resource in the
transition graph, check to see if it is missing the expire_attrs
meta-attribute or if it's set to false.  If so, clear the purge flag
from the remote's data.

When we get to remote_node_up, we'll skip purging the remote node's
attributes based on this.

Ref T133