Feature: controller: Copy remote resource uptime...
...from the peer's crm_node_t record to the remote's remote_ra_data_t.

When it comes time to check the uptime of the remote resource to
determine whether we should purge its transient attributes or not, we
only have access to the remote_ra_data_t.  We don't know anything about
the host that was previously managing the connection resource and its
uptime.

However, when we are processing the stop pseudo-action, we have all the
information we need.  We can look up the host that was managing the
stopping resource in its transition graph entry, we know the name of the
remote resource, and we already have the uptime stored in the peer node
cache.  So just look it up and copy the value on over.

Ref T133