Refactor: cts: Create a new LogKind enum.
Instead of relying on magic strings, we can use an enum to accomplish
the same thing.  The default is a local file, which has always been the
default if something other than "journal" or "remote" was given, so we
can get rid of various other magic strings that were being used.

Also, LogWatcher.__init__ can use a default value for the kind
parameter, which means we can enforce always having a default, which
means we probably don't need to check for None.