Refactor: libpacemaker, tools: Expect an enum to various messages.
In crm_mon, the crm-mon-disconnected message can just expect an enum
instead of an int.  We're already passing an enum and casting it when we
pop the arg off the stack.

In print_cluster_status, the cluster-status message is already expecting
an enum but our tools aren't smart enough and think the value being
passed is an int.  So just assign it to a variable with the right type
and pass that instead.