Refactor: python, cts: Test stats are now just a dict.
I originally imagined stats as their own module, with an enum limiting
what kinds of things could be stored as a statistic.  However, we use
them too widely and store various unpredictable things in them (search
for self.incr) to make this possible.

So instead, just make it a publically available dict and remove the
dict-like interface that was on CTSTest to access stats before.
Instead, you just use the stats directly.