Feature: daemons: Sync remote connection info with new controllers.
When a new controller connects to the cluster, include the current state
of what node hosts a remote connection as sub-elements of the acknak
message.  There could potentially be several remotes, so we need an
entire tree of elements instead of just a pair of attributes.

The new controller will iterate over the elements in this tree if they
are present and update its remote node cache.

Ref T133