Test: cts-scheduler: update tests for assignment change
Now that resources may be assigned to nodes with finite negative scores,
two scheduler regression tests have improvements:

In systemhealthp2, the stonith resource may now start, whereas before it was
left stopped. There are two nodes, hs21c (with health status yellow, equivalent
to -100 preference) and hs21d (which is unseen, so unclean and offline).
There was no good reason to leave stonith stopped.

In node-maintenance-1, rsc1 stays active where it is, whereas before it was
stopped. rsc1 is started on node1 (where it has a -1 location preference), and
rsc2 is started on node2 (where it has a -1 location preference), and node2 is
unmanaged. Since node2 is unmanaged, rsc2 can't be moved away from it, and rsc1
has nowhere to move to even though it has a negative preference for its current
node. Previously, rsc1 would be stopped even though it couldn't be recovered
elsewhere.