Test: cts-scheduler: update tests for colocation change
Now that mandatory colocations are considered first, the group-anticolocation
test is fixed.

Previously, the (failed) second member of group2 would be left stopped due to
its migration-threshold instead of moving the group to the other node, because
the optional anti-colocation of group1 and group2 was considered before
considering the mandatory implicit colocations of group members.

There are also minor score changes in one other test.