Low: daemons: pacemakerd -S should wait for shutdown before returning
Previously, `pacemakerd -S` would request a shutdown and then sit in a
loop for 30 minutes, waiting for the IPC connection to drop.  Only then
would it return to the command line.  This gave the user a way of
knowing the shutdown completed.

There are potentially various ways of solving this problem.  However, a
shutdown works by adding an event to the main loop which makes it
difficult to know exactly when it happens.  So, just go with the easiest
fix and re-implement the loop checking the IPC connection status.

Regression in 2.1.1 introduced by 64d4f9b5f