Refactor: controller: set_join_state needs to take a remote parameter.
Without this parameter, we won't know what to pass to as node_type to
cib__update_node_attr.  And without that, that function will not know to
update a remote node - it'll try to update a regular node by the same
name, which either doesn't exist or is not what we were hoping would
happen.

Ref T138