Feature: daemons: If it looks like attrd crashed, don't erase its attrs.
This portion runs on a peer that notices that some other node's attrd
has crashed.  If we can still reach the controller on that node via IPC,
attrd probably just segfaulted or otherwise abnormally terminated.
Don't erase its transient attributes so that when the node restarts
attrd, a peer can sync the previous values back over.

Ref T133