Feature: daemons: When attrd restarts, don't erase its attributes.
This is the companion piece to the previous patch.  It runs on the attrd
that crashed and restarted.  By default, it will erase all its transient
attributes.  When a peer syncs the values over, attrd will see these
attributes that it knows nothing about.  That will cause it to create
the attribute with a value of "", and it will then refuse to use the
synced value since the local value overrides it.

So, just check for PCMK_respawned in the environment and don't do the
local value override step if it's set.

Fixes T133