Fix: scheduler: ensure earlier group member starts happen after later member stops
Normally, group restart ordering is enforced by the group pseudo-actions:

  stop each member -> group is stopped -> start group -> start each member

However, in unusual circumstances (such as adding a new group member in the
middle of an active group with an unmanaged member), no member stops will be
scheduled, and thus the "group is stopped" pseudo-action won't exist. In that
situation, we want to block starts since all later members must be stopped
before a member can start.