Refactor: python, cts: Move FlipTest into the _cts module.
No significant code changes here.  I changed the import in CTStests.py
because it's rapidly going to get out of control.  On the other hand,
once all the tests are rearranged, I'm probably going to have to do
something totally different with how the list of tests is built which
might mean I won't need the import glob.