Refactor: python, cts: Support membership protocol in Environment.
Basically, just rename has_key to __contains__.  We can then use "xyz in
self._env" everywhere instead of making a list out of the keys and then
checking in that.

Also, simplify a block at the end of __getitem__.  We can use get() to
return a default value of None if the key is not present instead of
writing this exact same behavior by hand.