Refactor: based: New cib_op_attr_local value in enum cib_op_attr
This simplifies checking for primary and secondary requests in
parse_local_options_v2(), and it's extensible if more local-only
requests are added in the future.

This does have a side effect: PCMK__CIB_REQUEST_SECONDARY never had the
cib_op_attr_modifies flag (or its equivalent boolean in older code) set.
So we never matched it in the if condition. Now we do.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>