Test: scheduler: update expected graph for meta-attribute change
The comment before pcmk__insert_meta() in pcmk__create_migration_actions()
clearly indicates that record-pending=true is intended to be mandatory, but due
to the previous usage of add_hash_param(), the transition graph kept the
explicit op_defaults value of false. Now that it's corrected, update the
expected graph.