Refactor: based: Fix bad sync check in sync_our_cib()
Previously, if host was NULL, we would try to print it in the
crm_debug() call. Instead of just adding a pcmk__s() call, here we bail
out early and return an error if neither `all` nor `host` is set.

We could add a crm_log_xml_info() call for the request but it seems
unnecessary. Especially since the request itself would be fine if `all`
were set.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>