Refactor: libcib, based: Make cib_zero_copy work with more op types
Possibly "all op types," but hard to say...

If cib_zero_copy is set, then *result_cib == existing_cib in the op
processor functions. This causes problems for the processor functions
that free *result_cib and point it at a newly created CIB XML object.

The main part of the solution is to pass cib_perform_op() an
(xmlNode **) instead of (xmlNode *) for current_cib and adjust the
function body accordingly.

Ref T185

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>