Refactor: scheduler: drop redundant code in expression evaluation
If a node attribute expression is evaluated, there must be a node given, and
an attribute table is always created for nodes, so it will never be NULL when
it matters.

Besides, contrary to the commented intention, if there are no node attributes,
a test for whether an attribute is undefined can still pass.