Test: scheduler: add tests for ordering with unmanaged 'then'
These two tests cover the situation where there are two resources, one stopped
(but needing a start) and ordered before the other, which is unmanaged. There
are three variations: a symmetric "A then B" ordering, an asymmetric "A then B"
ordering, and an implicit ordering of group members.

For the unmanaged-ordering test, the unmanaged resource is started. In the
symmetric and group cases, the start of A should be blocked since B can't
restart. As of this commit, the symmetric case does not work.

For the unmanaged-ordering-stopped test, the unmanaged resource is stopped. In
all cases, the start of A should be allowed. As of this commit, all cases work.