Test: includes: Add pcmk__assert_exits.
This is very similar to pcmk__assert_asserts, except it checks that the
statement exits with the specified exit code.  This can be used for
testing things that call crm_exit.

Note that while the implementation of this is pretty close to
pcmk__assert_asserts, it's different enough to where I think it's worth
keeping them as separate macros.  Trying to do too much macro
refactoring makes it weird and hard to follow.