Build: devel: Filter directories out of the coverage report.
For now (and probably for a long time to come) there's no point in
generating coverage numbers for things in the daemons/ and tools/
directories.  For that matter, we never care about lib/gnu/ or replace/.

The impetus for this is that generating the man pages for these things
results in them being counted towards covered code, and I can't see any
obvious reason why this is the case.  Rather than spend more time
trying to fix it, I could just remove them entirely.

While I'm here, change from running "make core" to just "make".  All
this stuff ends up happening anyway.  We might as well control when.

Also, give the coverage report a title.