Test: scheduler: Clone instances should not shuffle unnecessarily
In some cases, clone instances may be shuffled when a new instance is
scheduled to start or promote. This can cause instances to be stopped
and started unnecessarily.

Here we add tests for three types of clones:
* "Bare" clones of primitives
* Clones of groups
* Clones of primitives within bundles (clone is bundle child resource)

For each clone type, we add four tests. In each test, no clone instance
is running on node 1, and a new instance should be started and possibly
promoted there.
* No constraints or stickiness
* Location constraint preferring node 1
* Promotable clone where node 1 has the highest promotion score
* Promotable clone where node 1 does not have the highest promotion
  score

The following tests are currently incorrect:
* clone-no-recover-shuffle-4 (shuffling)
* clone-no-recover-shuffle-5 (all instances of an anonymous clone move
  to one node)
* clone-no-recover-shuffle-6 (shuffling)
* clone-no-recover-shuffle-7 (shuffling)

Ref T489

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>